% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtrx.R
\name{mtrx_distance}
\alias{mtrx_distance}
\title{Function for generating report for distance matric.}
\usage{
mtrx_distance(
  a,
  filename = "",
  batch_size = 1000,
  metric = "kendall",
  type = "gpu"
)
}
\arguments{
\item{a}{Something.}

\item{filename}{CSV file.}

\item{batch_size}{int.}

\item{metric}{string for matric selection.}

\item{type}{"gpu" or "cpu".}
}
\value{
A list of correlation matrix, batch_a size and batch_b size.
}
\description{
Function for generating report for distance matric.
}
