% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtrx.R
\name{mtrx_read_kdm}
\alias{mtrx_read_kdm}
\title{get matrix from file.}
\usage{
mtrx_read_kdm(filename)
}
\arguments{
\item{filename}{File path for csv.}
}
\value{
data countmatrix data.
}
\description{
get matrix from file.
}
